+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Migrate to shared GA workflows. Refs UIPFPAT-63.
 * *BREAKING* migrate stripes dependencies to their Sunflower versions. Refs UIPFPAT-64.
 * *BREAKING* migrate react-intl to v7. Refs UIPFPAT-65.
+* React v19: refactor away from default props for functional components. Refs UIPFPAT-60.
 
 ## [6.2.0] (https://github.com/folio-org/ui-plugin-find-package-title/tree/v6.2.0) (2024-10-30)